NITTA.Frontends.Lua

Description

This module analyzes an abstract syntax tree of the Lua language source code, provided by Language.Lua module, and stores it into a NITTA's data flow graph.

Supported Lua costructions are:

Simple math operators (addition, subtraction, multiplication and division);

Variable assignments;

Bitwise left and right shifts;

Recursive calls.

The naming of variables in the output dataflow graph:

    x^1#2 -- for variables
    | | |
    | | +-- value access number (one for each), e.g.
    | |     x = 1; send(x); reg(x) -- two accesses
    | |
    | +---- value assignment number (one for each, optional)
    |       x = f(1); x = g(2) -- two assignments
    |
    +------ original variable name

    !123#3 -- for constant
      |  |
      |  +--- value access number
      |
      +------ value: 123

    _a#4 -- for an unnamed variable (example see below)
     | |
     | +--- value access number
     |
     +----- char of unnamed variable (a, b..., aa, ab, ...)

Example:

>>> :{
void $ mapM print $ functions (frDataFlow $ translateLua $ T.pack $ unlines
    [ "function f()"
    , "    local a = 1 + 2 + 3"
    , "    local b = a + 4 + 5"
    , "    b = b * 1 + 2"
    , "    c, d = b / 2"
    , "    send(b)"
    , "end"
    , "f()"
    ] :: DataFlowGraph String Int)
:}
const(1) = !1#0 = !1#1
const(2) = !2#0 = !2#1 = !2#2
!1#0 + !2#0 = _0#a
const(3) = !3#0
_0#a + !3#0 = a^0#0
const(4) = !4#0
a^0#0 + !4#0 = _0#b
const(5) = !5#0
_0#b + !5#0 = b^0#0
b^0#0 * !1#1 = _1#b
_1#b + !2#1 = b^1#0 = b^1#1
!2#2 / b^1#0 = _; !2#2 mod b^1#0 = _
send(b^1#1)

Documentation

translateLua :: (Var v, Val x) => Text -> FrontendResult v x

data FrontendResult v x

Constructors

FrontendResult

Fields

frDataFlow :: DataFlowGraph v x

frTrace :: [TraceVar]

frPrettyLog :: [HashMap v x] -> [HashMap String String]

data TraceVar

Constructors

TraceVar

Fields

tvFmt :: Maybe Text

tvVar :: Text

Internal

data LuaAlgBuilder x

Constructors

LuaAlgBuilder

Fields

algGraph :: [LuaStatement x]

A list containing all expressions to be added to the final graph.

algLatestLuaValueInstance :: HashMap Text LuaValueInstance

A table that maps a variable name to the most recent corresponding LuaValueInstance.

algVarCounters :: HashMap Text Int

A table needed to generate unique temporary variable names.

algVars :: HashMap LuaValueInstance [Text]

A table lists all uses of a particular LuaValueInstance.

algStartupArgs :: HashMap Int (Text, Text)

Map argument index to the variable name and initial value (in text).

algConstants :: HashMap Text LuaValueInstance

A table correlating constant with LuaValueInstance which store this constant.

algTraceFuncs :: [([Text], Maybe Text)]

A list that stores debug information about monitored variables and their display formats.

data LuaStatement x

Constructors

LuaStatement

Fields

fIn :: [Text]

fOut :: [LuaValueInstance]

fName :: Text

fValues :: [x]

fInt :: [Int]

data LuaValueInstance

Stores information about a particular version of a variable. The version of a variable changes after assigning a new value to it.

Constructors

LuaValueInstance

Fields

lviName :: Text

lviAssignCount :: Int

lviIsConstant :: Bool

findStartupFunction :: Block -> (Text, Stat, Stat)

getLuaBlockFromSources :: Text -> Block

processStatement :: (MonadState (LuaAlgBuilder x) m, Read x, Show x) => Text -> Stat -> m ()