NITTA.Intermediate.Analysis

Description

Much of the work in this module focuses on building waves of process execution. In this case, we call a wave a set of functions that are ready for execution (all input vars are ready to use) and that can be performed independently of each other. Example:

Lua code:

function sum(a)
    local d = a + 1
    sum(d)
end
sum(0)

After building DataFlowGraph, we get the following set of functions:

- const(1) = !1#0
- loop(0, d^0#0) = a^0#0
- a^0#0 + !1#0 = d^0#0

Const and Loop function can be executed at the first wave because all input variables are ready to use. After executing first wave a^0#0 and !1#0 will be ready, so we will be able to execute add function. So the resulting process waves are the following:

[
    ProcessWave {
        pwFs = [const(1) = !1#0,loop(0, d^0#0) = a^0#0],
        pwOut = fromList ["!10"]
    },
    ProcessWave {
        pwFs = [a^0#0 + !1#0 = d^0#0],
        pwOut = fromList ["d^0#0"]
    }
]

Documentation

data ProcessWave v x

Constructors

ProcessWave

Fields

pwFs :: [F v x]

Functions that can be executed at this wave

pwOut :: Set v

Set of output variables related to the functions from this step

buildProcessWaves :: (Var v, Val x) => [v] -> [F v x] -> [ProcessWave v x]

Divide function into execution waves.

reorderAlgorithm :: (Var v, Val x) => [F v x] -> [F v x]

Sort functions in order of execution. Note that some function could be executed in parallel, in this case we save order from the source list.

estimateVarWaves :: (Var v, Val x, Num a) => [v] -> [F v x] -> Map v a

Functions can be divided into waves of execution. For each output variable, we define the wave number on which the variable will be defined.