NITTA.Intermediate.Functions

Description

Library of functions for an intermediate algorithm representation. Execution relations between functions and process units are many-to-many.

function (functional block)

atomic operation in intermediate algorithm representation. Function has zero or many inputs and zero or many output. Function can contains state between process cycles.

Arithmetics

data Add v x

Constructors

Add (I v) (I v) (O v)

add :: (Var v, Val x) => v -> v -> [v] -> F v x

data Division v x

Constructors

Division

Fields

denom, numer :: I v

quotient, remain :: O v

division :: (Var v, Val x) => v -> v -> [v] -> [v] -> F v x

data Multiply v x

Constructors

Multiply (I v) (I v) (O v)

multiply :: (Var v, Val x) => v -> v -> [v] -> F v x

data ShiftLR v x

Functional unit that implements logic shift operations

Constructors

ShiftL Int (I v) (O v)

ShiftR Int (I v) (O v)

shiftL :: (Var v, Val x) => Int -> v -> [v] -> F v x

shiftR :: (Var v, Val x) => Int -> v -> [v] -> F v x

data Sub v x

Constructors

Sub (I v) (I v) (O v)

sub :: (Var v, Val x) => v -> v -> [v] -> F v x

data Neg v x

Constructors

Neg (I v) (O v)

neg :: (Var v, Val x) => v -> [v] -> F v x

module NITTA.Intermediate.Functions.Accum

Memory

data Constant v x

Constructors

Constant (X x) (O v)

constant :: (Var v, Val x) => x -> [v] -> F v x

isConst :: (Typeable v, Typeable x) => F v x -> Bool

data Loop v x

Loop -- function for transfer data between computational cycles. Let see the simple example with the following implementation of the Fibonacci algorithm.

Data flow graph:

    +---------------------------------+
    |                                 |
    v                                 |
+------+                          b2  |
| Loop |      b1_1  +-----+    +------+
+------+----+------>|     |    |
            | a1    | sum +----+
+------+----------->|     |
| Loop |    |       +-----+      b1_2
+------+    +-------------------------+
    ^                                 |
    |                                 |
    +---------------------------------+

Lua source code:

function fib(a1, b1)
    b2 = a1 + b1
    fib(b1, b2)
end
fib(0, 1)

Data flow defines computation for a single computational cycle. But a controller should repeat the algorithm infinite times, and usually, it is required to transfer data between cycles. Loop allows doing that. At first cycle, Loop function produces an initial value (`X x`), after that on each cycle Loop produces a variable value from the previous cycle, and consumes a new value at the end of the cycle.

Computational process:

         ][                 Cycle 1                 ][                Cycle 2                  ]
         ][                                         ][                                         ]
initial  ][ ---+                          b2   +--- ][ ---+                          b2   +--- ]
 value   ][ op |      b1_1  +-----+    +------>| Lo ][ op |      b1_1  +-----+    +------>| Lo ]
 is a    ][ ---+----+------>|     |    |       +--- ][ ---+----+------>|     |    |       +--- ]
part of  ][         |       | sum +----+            ][         |       | sum +----+            ]
software ][ ---+----------->|     |            +--- ][ ---+----------->|     |            +--- ]
         ][ op |    |       +-----+     b1_2   | Lo ][ op |    |       +-----+      b1_2  | Lo ]
         ][ ---+    +------------------------->+--- ][ ---+    +------------------------->+--- ]
         ][                                         ][                                         ]

Similation data:

Cycle number	a1	b1	b2

1	0	1	1

2	1	1	2

3	1	2	3

4	2	3	5

In practice, Loop function supported by Fram processor unit in the following way: Loop function should be prepared before execution by automatical refactor BreakLoop, which replace Loop by LoopEnd and LoopBegin.

Constructors

Loop (X x) (O v) (I v)

loop :: (Var v, Val x) => x -> v -> [v] -> F v x

isLoop :: (Typeable v, Typeable x) => F v x -> Bool

data LoopEnd v x

Constructors

LoopEnd (Loop v x) (I v)

data LoopBegin v x

Constructors

LoopBegin (Loop v x) (O v)

data Buffer v x

Constructors

Buffer (I v) (O v)

buffer :: (Var v, Val x) => v -> [v] -> F v x

Input/Output

newtype Receive v x

Constructors

Receive (O v)

receive :: (Var v, Val x) => [v] -> F v x

newtype Send v x

Constructors

Send (I v)

send :: (Var v, Val x) => v -> F v x

Internal

data BrokenBuffer v x

Special function for negative tests only.

Constructors

BrokenBuffer (I v) (O v)

brokenBuffer :: (Var v, Val x) => v -> [v] -> F v x

data Compare v x

Constructors

Compare CmpOp (I v) (I v) (O v)

data CmpOp

Constructors

CmpEq

CmpLt

CmpLte

CmpGt

CmpGte

cmp :: (Var v, Val x) => CmpOp -> v -> v -> [v] -> F v x

data TruthTable v x

Constructors

TruthTable (Map [Bool] Bool) [I v] (O v)

data LogicFunction v x

Constructors

LogicAnd (I v) (I v) (O v)

LogicOr (I v) (I v) (O v)

LogicNot (I v) (O v)

logicAnd :: (Var v, Val x) => v -> v -> [v] -> F v x

logicOr :: (Var v, Val x) => v -> v -> [v] -> F v x

logicNot :: (Var v, Val x) => v -> [v] -> F v x