NITTA.Intermediate.Functions.Accum

Description

Documentation

newtype Acc v x

Constructors

Acc

Fields

actions :: [Action v]

data Action v

Constructors

Push Sign (I v)

Pull (O v)

data Sign

Constructors

Plus

Minus

Acc to function

acc :: (Hashable v, Suffix v, ToString v, IsString v, Ord v, Num x, Typeable v, Typeable x) => [Action v] -> F v x

Create function with type F of Acc

accFromStr :: (Num x, Typeable x) => String -> F Text x

Special function for generating Acc from string, examples in tests

Utils functions

isPull :: Action v -> Bool

isPush :: Action v -> Bool