NITTA.Intermediate.Types

Description

Function interface

newtype I v

Input variable.

Constructors

I v

newtype O v

Output variable set.

Constructors

O (Set v)

newtype X x

Value of variable (constant or initial value).

Constructors

X x

Function description

data F v x where

Box forall functions.

Constructors

F

Fields

:: (Function f v, Patch f (v, v), Locks f v, Show f, Label f, FunctionSimulation f v x, Typeable f, Eq f)

=> { fun :: f

, funHistory :: [F v x]

} -> F v x

data FView

Helper for JSON serialization

Constructors

FView

Fields

fvFun :: Text

fvHistory :: [Text]

packF :: (Function f v, Patch f (v, v), Locks f v, Show f, Label f, FunctionSimulation f v x, Typeable f, Eq f) => f -> F v x

castF :: (Typeable f, Typeable v, Typeable x) => F v x -> Maybe (f v x)

Helper for extraction function from existential container F.

functionType :: F v x -> TypeRep

class Function f v | f -> v where

Type class for application algorithm functions.

Minimal complete definition

Nothing

Methods

inputs :: f -> Set v

Get all input variables.

outputs :: f -> Set v

Get all output variables.

isInternalLockPossible :: f -> Bool

Sometimes, one function can cause internal process unit lock for another function.

data Lock v

Variable casuality.

Constructors

Lock

Fields

locked :: v

lockBy :: v

class Var v => Locks x v | x -> v where

Casuality of variable processing sequence in term of locks.

For example: > c := a + b > [ Lock{ locked=c, lockBy=a }, Lock{ locked=c, lockBy=b } ]

Methods

locks :: x -> [Lock v]

inputsLockOutputs :: Function f v => f -> [Lock v]

All input variables locks all output variables.

class WithFunctions a f | a -> f where

Type class of something, which is related to functions.

Methods

functions :: a -> [f]

Get a list of associated functions.

class Label a where

Type class for making fine label for Functions.

Methods

label :: a -> String

Functional simulation

class FunctionSimulation f v x | f -> v x where

The type class for function simulation.

Methods

simulate :: CycleCntx v x -> f -> [(v, x)]

Receive a computational context and return changes (list of varible names and its new values).

newtype CycleCntx v x

Constructors

CycleCntx

Fields

cycleCntx :: HashMap v x

data Cntx v x

Constructors

Cntx

Fields

cntxProcess :: [CycleCntx v x]

all variables on each process cycle

cntxReceived :: Map v [x]

sequences of all received values, one value per process cycle

cntxCycleNumber :: Int

log2md :: ToString a => [HashMap a String] -> String

>>> let records = map HM.fromList [[("x1"::String,"1.2"::String), ("x2","3.4")], [("x1","3.4"), ("x2","2.3")]]
>>> putStr $ log2md records
| Cycle  | x1   | x2   |
|:-------|:-----|:-----|
| 1      | 1.2  | 3.4  |
| 2      | 3.4  | 2.3  |

log2json :: ToString a => [HashMap a String] -> ByteString

>>> import qualified Data.ByteString.Lazy.Char8 as BS
>>> let records = map HM.fromList [[("x1"::String,"1.2"::String), ("x2","3.4")], [("x1","3.4"), ("x2","2.3")]]
>>> BS.putStr $ log2json records
[
    {
        "x1": 1.2,
        "x2": 3.4
    },
    {
        "x1": 3.4,
        "x2": 2.3
    }
]

log2csv :: ToString a => [HashMap a String] -> ByteString

>>> import qualified Data.ByteString.Lazy.Char8 as BS
>>> let records = map HM.fromList [[("x1"::String,"1.2"::String), ("x2","3.4")], [("x1","3.4"), ("x2","2.3")]]
>>> BS.putStr $ log2csv records
x1,x2
1.2,3.4
3.4,2.3

cntxReceivedBySlice :: Ord v => Cntx v x -> [Map v x]

Make sequence of received values '[ Map v x ]'

getCntx :: (Hashable a, ToString a) => CycleCntx a v -> a -> v

updateCntx :: (Hashable v, ToString v) => CycleCntx v x -> [(v, x)] -> Either String (CycleCntx v x)

Patch

class Patch f diff where

Patch class allows replacing one variable by another. Especially for algorithm refactor.

Methods

patch :: diff -> f -> f

data Changeset v

Change set for patch.

>>> Changeset (M.fromList [("a", "b"), ("c", "d")]) (M.fromList [("e", S.fromList ["f", "g"])]) :: Changeset String
Changeset{changeI=[(a, b), (c, d)], changeO=[(e, [f, g])]}

Constructors

Changeset

Fields

changeI :: Map v v

change set for input variables (one to one)

changeO :: Map v (Set v)

change set for output variables. Many to many relations:

fromList [(a, {x}), (b, {x})] -- several output variables to one
fromList [(c, {y, z})] -- one output variable to many

reverseDiff :: Ord t => Changeset t -> Changeset t

Reverse changeset for patch a process unit options / decision.

module NITTA.Intermediate.Value

module NITTA.Intermediate.Variable