NITTA.Intermediate.Value

Description

Type classes

class (Typeable x, Show x, Read x, PrintfArg x, Default x, Integral x, Enum x, Eq x, Num x, Bits x, Validity x, FixedPointCompatible x, ToJSON x) => Val x where

Type class for representation of processing values. A value should include two parts: data and attribute.

Minimal complete definition

dataWidth, rawData, rawAttr, fromRaw, dataLiteral

Methods

dataWidth :: x -> Int

data bus width

attrWidth :: x -> Int

attribute bus width

rawData :: x -> Integer

raw representation of data (Integer)

rawAttr :: x -> Integer

raw representation of attributes (Integer)

fromRaw :: Integer -> Integer -> x

construct a value from raw data and attributes

dataLiteral :: x -> Text

convert a value to Verilog literal with data

attrLiteral :: x -> Text

convert a value to Verilog literal with attributes

verilogHelper :: x -> Text

helper functions to work with values in Verilog (trace and assert)

verilogAssertRE :: x -> Regex

RE for extraction assertion data from a testbench log

class Default x => DefaultX u x | u -> x where

Minimal complete definition

Nothing

Methods

defX :: u -> x

class FixedPointCompatible a where

Type class for values, which contain information about fractional part of value (for fixed point arithmetics).

Methods

scalingFactorPower :: a -> Integer

fractionalBitSize :: a -> Int

scalingFactor :: (Floating a, FixedPointCompatible a) => a -> a

minMaxRaw :: (Num b, Val p) => p -> (b, b)

Minimal and maximal raw value.

Compositional type

data Attr x

All values with attributes.

Constructors

Attr

Fields

value :: x

invalid :: Bool

Value types

newtype IntX (w :: Nat)

Integer number with specific bit width.

Constructors

IntX

Fields

intX :: Integer

newtype FX (m :: Nat) (b :: Nat)

Number with fixed point. FX m b where

m the number of magnitude or integer bits

b the total number of bits

fxm.b: The "fx" prefix is similar to the above, but uses the word length as the second item in the dotted pair. For example, fx1.16 describes a number with 1 magnitude bit and 15 fractional bits in a 16 bit word.

Constructors

FX

Fields

rawFX :: Integer