NITTA.LuaFrontend

Description

This module analyzes an abstract syntax tree of the Lua language source code, provided by Language.Lua module, and stores it into a NITTA's data flow graph.

Supported Lua costructions are:

Simple math operators (addition, subtraction, multiplication and division);

Variable assignments;

Bitwise left and right shifts;

Recursive calls.

Documentation

lua2functions :: (ToString a, Suffix a, Hashable a, Val t, IsString a, Ord a, Typeable a) => Text -> FrontendResult a t

data FrontendResult v x

Constructors

FrontendResult

Fields

frDataFlow :: DataFlowGraph v x

frTrace :: [TraceVar]

frPrettyLog :: [HashMap v x] -> [HashMap String String]

data TraceVar

Constructors

TraceVar

Fields

tvFmt, tvVar :: Text

Internal

data LuaAlgBuilder x

Constructors

LuaAlgBuilder

Fields

algGraph :: [LuaStatement x]

A list containing all expressions to be added to the final graph.

algLatestLuaValueInstance :: HashMap Text LuaValueInstance

A table that maps a variable name to the most recent corresponding LuaValueInstance.

algVarCounters :: HashMap Text Int

A table needed to generate unique temporary variable names.

algVars :: HashMap LuaValueInstance [Text]

A table lists all uses of a particular LuaValueInstance.

algStartupArgs :: HashMap Int (Text, Text)

A table correlating the ordinal number of an argument with a variable storing its value and startup value of this variable.

algConstants :: HashMap Text LuaValueInstance

A table correlating constant with LuaValueInstance which store this constant.

algTraceFuncs :: [([Text], Text)]

A list that stores debug information about monitored variables and their display formats.

data LuaStatement x

Constructors

LuaStatement

Fields

fIn :: [Text]

fOut :: [LuaValueInstance]

fName :: Text

fValues :: [x]

fInt :: [Int]

data LuaValueInstance

Stores information about a particular version of a variable. The version of a variable changes after assigning a new value to it.

Constructors

LuaValueInstance

Fields

lviName :: Text

lviAssignCount :: Int

lviIsConstant :: Bool

findStartupFunction :: Block -> (Text, Stat, Stat)

getLuaBlockFromSources :: Text -> Block

processStatement :: (MonadState (LuaAlgBuilder x) m, Read x, Show x) => Text -> Stat -> m ()