NITTA.Model.Networks.Bus

Description

Documentation

data BusNetwork tag v x t

Constructors

BusNetwork

Fields

bnName :: tag

bnRemains :: [F v x]

List of functions bound to network, but not bound to any process unit.

bnBound :: Map tag [F v x]

Map process unit name to list of bound functions.

bnProcess :: Process t (StepInfo v x t)

Network process (bindings and transport instructions)

bnPus :: Map tag (PU v x t)

Map of process units.

bnSignalBusWidth :: Int

Controll bus width.

ioSync :: IOSynchronization

bnEnv :: UnitEnv (BusNetwork tag v x t)

bnPUPrototypes :: Map tag (PUPrototype tag v x t)

Set of the PUs that could be added to the network during synthesis process

data family Instruction pu :: Type

data family Ports pu :: Type

A processor unit control ports (signals, flags).

data family IOPorts pu :: Type

boundFunctions :: Ord k => k -> BusNetwork k v x t -> [F v x]

controlSignalLiteral :: (Builder (IsCustomSink dst) dst ~ Builder (IsCustomSink dst) dst, Interpolatable (IsCustomSink dst) src dst) => src -> dst

busNetwork :: Default t => tag -> IOSynchronization -> BusNetwork tag v x t

Builder

modifyNetwork :: BusNetwork k v x t -> State (BuilderSt k v x t) a -> BusNetwork k v x t

defineNetwork :: Default t => k -> IOSynchronization -> State (BuilderSt k v x t) a -> BusNetwork k v x t

addCustom :: forall tag v x t m pu. (MonadState (BuilderSt tag v x t) m, PUClasses pu v x t, UnitTag tag) => tag -> pu -> IOPorts pu -> m ()

add :: (MonadState (BuilderSt tag v x t) m, PUClasses pu v x t, Default pu, UnitTag tag) => tag -> IOPorts pu -> m ()

Add PU with the default initial state. Type specify by IOPorts.

addPrototype :: (MonadState (BuilderSt tag v x t) m, PUClasses pu v x t, Default pu, UnitTag tag) => tag -> IOPorts pu -> m ()

Add PU to prototypes with the default initial state. Type specify by IOPorts.

addCustomPrototype :: forall tag v x t m pu. (MonadState (BuilderSt tag v x t) m, PUClasses pu v x t, UnitTag tag) => tag -> pu -> IOPorts pu -> m ()