NITTA.Model.Networks.Types

Description

Documentation

data PU v x t where

Existential container for a processor unit .

Constructors

PU

Fields

:: PUClasses pu v x t

=> { unit :: pu

, diff :: Changeset v

, uEnv :: UnitEnv pu

} -> PU v x t

unitType :: PU v x t -> TypeRep

type PUClasses pu v x t = (ByTime pu t, Connected pu, IOConnected pu, EndpointProblem pu v t, BreakLoopProblem pu v x, ConstantFoldingProblem pu v x, OptimizeAccumProblem pu v x, OptimizeLogicalUnitProblem pu v x, ResolveDeadlockProblem pu v x, ProcessorUnit pu v x t, Show (Instruction pu), Typeable pu, UnambiguouslyDecode pu, TargetSystemComponent pu, Controllable pu, IOTestBench pu v x, Locks pu v, Typeable pu)

data IOSynchronization

Constructors

Sync
IO cycle synchronously to process cycle

ASync
if IO cycle lag behiend - ignore them

OnBoard
defined by onboard signal (sync - false, async - true)

data PUPrototype tag v x t where

PU and some additional information required for allocation on BusNetwork

Constructors

PUPrototype

Fields

:: (UnitTag tag, PUClasses pu v x t)

=> { pTag :: tag

Prototype tag. You can specify tag as a template by adding {x}. This will allow to allocate PU more than once by replacing {x} with index. When PU is allocated processUnitTag will look like bnName_pTag.

, pProto :: pu

PU prototype

, pIOPorts :: IOPorts pu

IO ports that will be used by PU

} -> PUPrototype tag v x t

puInputPorts :: PU v x t -> Set InputPortTag

puOutputPorts :: PU v x t -> Set OutputPortTag

puInOutPorts :: PU v x t -> Set InoutPortTag