NITTA.Model.Problems.Dataflow

Description

Documentation

data DataflowSt tag v tp

Dataflow option (tp ~ TimeConstraint t) or decision (tp Z Interval t) statement. Describe sending data between processor units over a network. Any DataflowSt has implicently linked NITTA.Model.Problems.Endpoint.

Constructors

DataflowSt

Fields

dfSource :: (tag, EndpointSt v tp)

A source processor unit of data flow transaction, and it's time constrains which defines when data can be sended.

dfTargets :: [(tag, EndpointSt v tp)]

All possible targets of dataflow transaction.

class DataflowProblem u tag v t | u -> tag v t where

Implemented for any things, which can send data between processor units over the network.

Methods

dataflowOptions :: u -> [DataflowSt tag v (TimeConstraint t)]

dataflowDecision :: u -> DataflowSt tag v (Interval t) -> u

dataflowOption2decision :: Time t => DataflowSt tag v (TimeConstraint t) -> DataflowSt tag v (Interval t)

Convert dataflow option to decision.