NITTA.Model.Problems.Endpoint

Description

Documentation

data EndpointSt v tp

Constructors

EndpointSt

Fields

epRole :: EndpointRole v

use processor unit as source or target of data

epAt :: tp

time of operation

class EndpointProblem u v t | u -> v t where

Methods

endpointOptions :: u -> [EndpointSt v (TimeConstraint t)]

endpointDecision :: u -> EndpointSt v (Interval t) -> u

data EndpointRole v

Constructors

Source (Set v)
get data from PU

Target v
put data to PU

endpointOptionToDecision :: (Num a, Ord a) => EndpointSt v (TimeConstraint a) -> EndpointSt v (Interval a)

The simplest way to convert an endpoint synthesis option to a endpoint decision.

isSource :: EndpointSt v tp -> Bool

isTarget :: EndpointSt v tp -> Bool

isSubroleOf :: Ord a => EndpointRole a -> EndpointRole a -> Bool

setAt :: tp -> EndpointSt v tp -> EndpointSt v tp

updAt :: (t -> tp) -> EndpointSt v t -> EndpointSt v tp