NITTA.Model.Problems.Refactor.BreakLoop

Description

BreakLoop example:

BreakLoop x o i

before:

l@( Loop (X x) (O o) (I i) )

after:

LoopEnd l (I i)
LoopBegin l (O o)

For more details: Loop

Documentation

data BreakLoop v x

Constructors

BreakLoop

Fields

loopX :: x

initial looped value

loopO :: Set v

output variables

loopI :: v

input variable

class BreakLoopProblem u v x | u -> v x where

Minimal complete definition

Nothing

Methods

breakLoopOptions :: u -> [BreakLoop v x]

breakLoopDecision :: u -> BreakLoop v x -> u

recLoop :: (Hashable v, Suffix v, ToString v, IsString v, Ord v, Show x, Typeable v, Typeable x, Eq x) => BreakLoop v x -> F v x

recLoopIn :: (Hashable v, Suffix v, ToString v, IsString v, Ord v, Show x, Typeable v, Typeable x, Eq x) => BreakLoop v x -> F v x

recLoopOut :: (Hashable v, Suffix v, ToString v, IsString v, Ord v, Show x, Typeable v, Typeable x, Eq x) => BreakLoop v x -> F v x