NITTA.Model.Problems.Refactor.ResolveDeadlock

Description

Possible deadlocks (recognized in Bus):

selfsending;

                    source1
    DoubleLoopOut--b1--+--b2--> DoubleLoopIn
               |       +------> Send
               |
               +------b3------> DoubleLoopIn
                    source2

    Possible buffers: b1, b2, b3

classical deadlock betweeb two function on same PU.

    a + b = c ---+----> c + d = e ---> e * c = f
                 |                         ^
                 +-------------b1----------+

ResolveDeadlock example:

ResolveDeadlock [a, b]

before:

f1 :: (...) -> ([a, b])
f2 :: (a, ...) -> (...)
f3 :: (b, ...) -> (...)

f1, f2 and f3 process on same process unit. In this case, we have deadlock, which can be fixed by insertion of buffer between functions.

after:

f1 :: (...) -> ([a@buf])
buffer :: a@buf -> ([a, b])
f2 :: (a, ...) -> (...)
f3 :: (b, ...) -> (...)

Documentation

data ResolveDeadlock v x

Constructors

ResolveDeadlock

Fields

newBuffer :: F v x

changeset :: Changeset v

class ResolveDeadlockProblem u v x | u -> v x where

Minimal complete definition

Nothing

Methods

resolveDeadlockOptions :: u -> [ResolveDeadlock v x]

resolveDeadlockDecision :: u -> ResolveDeadlock v x -> u

resolveDeadlock :: (Var v, Val x) => Set v -> ResolveDeadlock v x

maxBufferStack :: Int

The constant, which restrict maximum length of a buffer sequence.