NITTA.Model.ProcessorUnits.Multiplier

Description

Processor unit

A processor unit (PU) can be used for:

data storage and processing;

interaction with the periphery (IO);

control of a NITTA processor (special case).

There are characterized by complicated behavior with:

multifunctionality;

internal parallelism;

superscalar;

pipelining;

availability of internal resources.

The multiplier PU is one of the simplest processors because it realizes only one function with sequence evaluation (Multiply). Processor behavior in a specific application is determined by the applied algorithm (DataFlow).

Any PU may include three components:

hardware - set of prepared or automatically generated hardware descriptions (/hdl/multiplier);

software - set of binary files that determine:

initial state and setting ;

a control program;

PU model - CAD component that implements PU support (hardware and software generation, instance generation, computation process scheduling, testing environment, etc.).

All three components are hardly related to each other and needed to comply with each other strictly. For a deeper understanding, a PU developer should understand all of its components. The multiplier model will be described above.

Processor unit model

A model purpose is "teaching" CAD how to work with the PU:

Which functions can be evaluated by PU? (see ProcessorUnit)?

How to control PU for evaluating specific functions (see ProcessorUnit, Controllable)?

How to translating instructions to microcode (see UnambiguouslyDecode)?

What are the options of PU synthesis decision available? (see ProcessorUnit, EndpointDT):

push variable to the PU (Target);

pull at least one variable from the PU (Source).

The basis of a PU model is a data structure that represents:

PU state during computation process scheduling;

process description (full or fragment), which can be translated to microcode.

Exactly around this data structure, all algorithmic part of the PU model is developed. The types of the following variables parametrize the data structure:

v - variable id (usually String);

x - a type of processed value (see Val);

t - time moment id (usually Int).

Multiplier processor unit

The multiplier processor unit can evaluate the following function type:

Multiply.

Only one function can be processed in one moment, and its execution cannot be interrupted.

This module should be considered as a tutorial for the development of other models of processor units. Its source code is written almost in literature style, so we recommend to continue reading within the source code.

Interaction with multiplier processor unit

We will consider the example of the computation process scheduling for one function. To do this, we need to start GHCi interpreter by executing `stack repl` command from the project directory. After that:

> :l NITTA.Model.ProcessorUnits.Multiplier
[ 1 of 30] Compiling NITTA.Intermediate.Value ( UserspenskoiDocumentsnitta-corpnittasrcNITTAIntermediate/Value.hs, interpreted )
[ 2 of 30] Compiling NITTA.Intermediate.Variable ( UserspenskoiDocumentsnitta-corpnittasrcNITTAIntermediate/Variable.hs, interpreted )
[ 3 of 30] Compiling NITTA.Intermediate.Types ( UserspenskoiDocumentsnitta-corpnittasrcNITTAIntermediate/Types.hs, interpreted )
[ 4 of 30] Compiling NITTA.Model.Problems.Binding ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProblemsBinding.hs, interpreted )
[ 5 of 30] Compiling NITTA.Model.Types ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModel/Types.hs, interpreted )
[ 6 of 30] Compiling NITTA.Model.Problems.Endpoint ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProblemsEndpoint.hs, interpreted )
[ 7 of 30] Compiling NITTA.Model.Problems.Dataflow ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProblemsDataflow.hs, interpreted )
[ 8 of 30] Compiling NITTA.Project.Types ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProject/Types.hs, interpreted )
[ 9 of 30] Compiling NITTA.Utils.Base ( UserspenskoiDocumentsnitta-corpnittasrcNITTAUtils/Base.hs, interpreted )
[10 of 30] Compiling NITTA.Intermediate.Functions.Accum ( UserspenskoiDocumentsnitta-corpnittasrcNITTAIntermediateFunctionsAccum.hs, interpreted )
[11 of 30] Compiling NITTA.Intermediate.Functions ( UserspenskoiDocumentsnitta-corpnittasrcNITTAIntermediate/Functions.hs, interpreted )
[12 of 30] Compiling NITTA.Model.Problems.Refactor ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProblemsRefactor.hs, interpreted )
[13 of 30] Compiling NITTA.Model.Problems.Whole ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProblemsWhole.hs, interpreted )
[14 of 30] Compiling NITTA.Model.Problems ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModel/Problems.hs, interpreted )
[15 of 30] Compiling NITTA.Utils.CodeFormat ( UserspenskoiDocumentsnitta-corpnittasrcNITTAUtils/CodeFormat.hs, interpreted )
[16 of 30] Compiling NITTA.Model.ProcessorUnits.Types ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProcessorUnitsTypes.hs, interpreted )
[17 of 30] Compiling NITTA.Utils      ( UserspenskoiDocumentsnitta-corpnittasrcNITTAUtils.hs, interpreted )
[18 of 30] Compiling NITTA.Project.VerilogSnippets ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProject/Snippets.hs, interpreted )
[19 of 30] Compiling NITTA.Project.Implementation ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProject/Implementation.hs, interpreted )
[20 of 30] Compiling NITTA.Project.Parts.Utils ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProjectPartsUtils.hs, interpreted )
[21 of 30] Compiling NITTA.Project.TestBench ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProjectPartsTestBench.hs, interpreted )
[22 of 30] Compiling NITTA.Project.Parts.TargetSystem ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProjectPartsTargetSystem.hs, interpreted )
[23 of 30] Compiling NITTA.Project.Parts.Icarus ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProjectPartsIcarus.hs, interpreted )
[24 of 30] Compiling NITTA.Model.Networks.Types ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelNetworksTypes.hs, interpreted )
[25 of 30] Compiling NITTA.Utils.ProcessDescription ( UserspenskoiDocumentsnitta-corpnittasrcNITTAUtils/ProcessDescription.hs, interpreted )
[26 of 30] Compiling NITTA.Model.Networks.Bus ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelNetworksBus.hs, interpreted )
[27 of 30] Compiling NITTA.Project.Parts.Quartus ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProjectPartsQuartus.hs, interpreted )
[28 of 30] Compiling NITTA.Project.Utils ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProject/Utils.hs, interpreted )
[29 of 30] Compiling NITTA.Project    ( UserspenskoiDocumentsnitta-corpnittasrcNITTAProject.hs, interpreted )
[30 of 30] Compiling NITTA.Model.ProcessorUnits.Multiplier ( UserspenskoiDocumentsnitta-corpnittasrcNITTAModelProcessorUnitsMultiplier.hs, interpreted )
Ok, 30 modules loaded.
> :module +NITTA.Model.ProcessorUnits.Types NITTA.Intermediate.Functions Numeric.Interval.NonEmpty Data.Set Prettyprinter.Render.Text
> :set prompt "ESC[34mλ> ESC[m"

Now create the function and multiplier model initial state. Unfortunately, it is not enough information for GHC deduction of its type, so let's define its implicitly.

>>> :module +Prettyprinter.Render.Text
>>> let f = F.multiply "a" "b" ["c", "d"] :: F String Int
>>> f
a * b = c = d
>>> let st0 = multiplier True :: Multiplier String Int Int
>>> putDoc $ pretty st0
Multiplier:
    remain: []
    targets: []
    sources: []
    currentWork: Nothing
    isMocked: True
    Process:
        steps:
        relations:
        nextTick: 0
        nextUid: 0
>>> endpointOptions st0
[]

Bind a function to the multiplier unit. This operation could be executed at any time of working with a model, including when a computation process is fully scheduled (new work can be added). The main rules are: 1) if work is fully scheduled, then it is necessary to perform it, and any part of it cannot be "lost" inside the model; 2) if a unit has its internal resources, there should be enough to finish schedule, even it is inefficient.

>>> let Right st1 = tryBind f st0
>>> putDoc $ pretty st1
Multiplier:
    remain: [a * b = c = d]
    targets: []
    sources: []
    currentWork: Nothing
    isMocked: True
    Process:
        steps:
        relations:
        nextTick: 0
        nextUid: 0
>>> endpointOptions st1
[?Target a@(0..INF /P 1..INF),?Target b@(0..INF /P 1..INF)]

As we can see, after binding, we have two different options of computational process scheduling that match different argument loading sequences: a or b. We can see that they are similar from an execution sequence point of view: loading can be started from 0 tick or after an arbitrary delay; for loading of one argument needed only one tick, but it can continue for an arbitrary time. Choose the variant.

>>> let st2 = endpointDecision st1 $ EndpointSt (Target "a") (0...2)
>>> putDoc $ pretty st2
Multiplier:
    remain: []
    targets: ["b"]
    sources: ["c","d"]
    currentWork: Just a * b = c = d
    isMocked: True
    Process:
        steps:
            0) Step {pID = 0, pInterval = 0 ... 2, pDesc = Endpoint: Target a}
            1) Step {pID = 1, pInterval = 0 ... 2, pDesc = Instruction: Load}
        relations:
            0) Vertical {vUp = 0, vDown = 1}
        nextTick: 3
        nextUid: 2
>>> mapM_ print $ endpointOptions st2
?Target b@(3..INF /P 1..INF)
>>> let st3 = endpointDecision st2 $ EndpointSt (Target "b") (3...3)
>>> putDoc $ pretty st3
Multiplier:
    remain: []
    targets: []
    sources: ["c","d"]
    currentWork: Just a * b = c = d
    isMocked: True
    Process:
        steps:
            0) Step {pID = 0, pInterval = 0 ... 2, pDesc = Endpoint: Target a}
            1) Step {pID = 1, pInterval = 0 ... 2, pDesc = Instruction: Load}
            2) Step {pID = 2, pInterval = 3 ... 3, pDesc = Endpoint: Target b}
            3) Step {pID = 3, pInterval = 3 ... 3, pDesc = Instruction: Load}
        relations:
            0) Vertical {vUp = 2, vDown = 3}
            1) Vertical {vUp = 0, vDown = 1}
        nextTick: 4
        nextUid: 4
>>> mapM_ print $ endpointOptions st3
?Source c,d@(6..INF /P 1..INF)

After loading both arguments, we can see that the next option is unloading c and d variables. Note, these variables can be unloaded either concurrently or sequentially (for details, see how the multiplier works inside). Consider the second option:

>>> let st4 = endpointDecision st3 $ EndpointSt (Source $ S.fromList ["c"]) (6...6)
>>> putDoc $ pretty st4
Multiplier:
    remain: []
    targets: []
    sources: ["d"]
    currentWork: Just a * b = c = d
    isMocked: True
    Process:
        steps:
            0) Step {pID = 0, pInterval = 0 ... 2, pDesc = Endpoint: Target a}
            1) Step {pID = 1, pInterval = 0 ... 2, pDesc = Instruction: Load}
            2) Step {pID = 2, pInterval = 3 ... 3, pDesc = Endpoint: Target b}
            3) Step {pID = 3, pInterval = 3 ... 3, pDesc = Instruction: Load}
            4) Step {pID = 4, pInterval = 6 ... 6, pDesc = Endpoint: Source c}
            5) Step {pID = 5, pInterval = 6 ... 6, pDesc = Instruction: Out}
        relations:
            0) Vertical {vUp = 4, vDown = 5}
            1) Vertical {vUp = 2, vDown = 3}
            2) Vertical {vUp = 0, vDown = 1}
        nextTick: 7
        nextUid: 6
>>> mapM_ print $ endpointOptions st4
?Source d@(7..INF /P 1..INF)
>>> let st5 = endpointDecision st4 $ EndpointSt (Source $ S.fromList ["d"]) (7...7)
>>> putDoc $ pretty st5
Multiplier:
    remain: []
    targets: []
    sources: []
    currentWork: Nothing
    isMocked: True
    Process:
        steps:
            0) Step {pID = 0, pInterval = 0 ... 2, pDesc = Endpoint: Target a}
            1) Step {pID = 1, pInterval = 0 ... 2, pDesc = Instruction: Load}
            2) Step {pID = 2, pInterval = 3 ... 3, pDesc = Endpoint: Target b}
            3) Step {pID = 3, pInterval = 3 ... 3, pDesc = Instruction: Load}
            4) Step {pID = 4, pInterval = 6 ... 6, pDesc = Endpoint: Source c}
            5) Step {pID = 5, pInterval = 6 ... 6, pDesc = Instruction: Out}
            6) Step {pID = 6, pInterval = 7 ... 7, pDesc = Endpoint: Source d}
            7) Step {pID = 7, pInterval = 7 ... 7, pDesc = Instruction: Out}
            8) Step {pID = 8, pInterval = 0 ... 7, pDesc = Intermediate: a * b = c = d}
        relations:
            0) Vertical {vUp = 8, vDown = 6}
            1) Vertical {vUp = 8, vDown = 4}
            2) Vertical {vUp = 8, vDown = 2}
            3) Vertical {vUp = 8, vDown = 0}
            4) Vertical {vUp = 6, vDown = 7}
            5) Vertical {vUp = 4, vDown = 5}
            6) Vertical {vUp = 2, vDown = 3}
            7) Vertical {vUp = 0, vDown = 1}
        nextTick: 8
        nextUid: 9
>>> endpointOptions st5
[]

All options of computing process scheduling are run out. All bound functions are planned. Further microcode can be generated, which can be organizing the described computational process on the multiplier.

Documentation

multiplier :: Default t => Bool -> Multiplier v x t

Multiplier PU model constructor. Argument defines the computation unit's internal organization: using multiplier IP kernel (False) or mock (True). For more information, look hardware function in TargetSystemComponent class.

data Multiplier v x t

It is a PU model state representation, which describes each state of synthesis model for that PU.

data family Ports pu :: Type

A processor unit control ports (signals, flags).

data family IOPorts pu :: Type