NITTA.Model.ProcessorUnits.Types

Description

Processor unit

class (Typeable tag, Ord tag, ToString tag, IsString tag, Semigroup tag) => UnitTag tag where

Class for processor unit tag or "name"

Methods

isTemplate :: tag -> Bool

Whether the value can be used as a template or not

fromTemplate :: tag -> String -> tag

Create tag from the template and index

class VarValTime v x t => ProcessorUnit u v x t | u -> v x t where

Process unit - part of NITTA process with can execute a function from intermediate representation:

get function for execution (tryBind);

store computational process description (process);

other features implemented by different type classes (see above and in NITTA.Model.Problems).

Minimal complete definition

tryBind, process

Methods

tryBind :: F v x -> u -> Either String u

process :: u -> Process t (StepInfo v x t)

parallelismType :: u -> ParallelismType

Indicates what type of parallelism is supported by ProcessorUnit

puSize :: u -> Float

Provide the processor unit size. At the moment it's just the number of subprocessors

bind :: ProcessorUnit u v x t => F v x -> u -> u

allowToProcess :: ProcessorUnit b v x t => F v x -> b -> Bool

class NextTick u t | u -> t where

Methods

nextTick :: u -> t

data ParallelismType

Processor unit parallelism type

Constructors

Full
All operations can be performed in parallel mode

Pipeline
All operations can be performed in pipeline mode

None
Other processor units

Process description

data Process t i

Computational process description. It was designed in ISO 15926 style, with separated data and relations storage.

Constructors

Process

Fields

steps :: [Step t i]

All process steps desctiption.

relations :: [Relation]

List of relationships between process steps (see Relation).

nextTick_ :: t

Next tick for instruction. Note: instruction /= endpoint.

nextUid :: ProcessStepID

Next process step ID

type ProcessStepID = Int

Unique ID of a process step. Uniquity presented only inside PU.

data Step t i

Process step representation

Constructors

Step

Fields

pID :: ProcessStepID

uniq (inside single the process unit) step ID

pInterval :: Interval t

step time

pDesc :: i

step description

data StepInfo v x t where

Informative process step description at a specific process level.

Constructors

CADStep :: String -> StepInfo v x t
CAD level step

RefactorStep :: (Typeable ref, Show ref, Eq ref) => ref -> StepInfo v x t
Apply refactoring

IntermediateStep :: F v x -> StepInfo v x t
intermidiate level step (function execution)

EndpointRoleStep :: EndpointRole v -> StepInfo v x t
endpoint level step (source or target)

InstructionStep :: (Show (Instruction pu), Typeable (Instruction pu)) => Instruction pu -> StepInfo v x t
process unit instruction (depends on process unit type)

NestedStep
wrapper for nested process unit step (used for networks)

Fields

:: UnitTag tag

=> { nTitle :: tag

, nStep :: Step t (StepInfo v x t)

} -> StepInfo v x t

AllocationStep :: (Typeable a, Show a, Eq a) => a -> StepInfo v x t
Process unit allocation step

data Relation

Relations between process steps.

Constructors

Vertical
Vertical relationships (up and down). For example, the intermediate step (function execution) can be translated to a sequence of endpoint steps (receiving and sending variable), and process unit instructions.

Fields

vUp, vDown :: ProcessStepID

Horizontal
Horizontal relationships (on one level). For example, we bind the function and apply the refactoring. The binding step should be connected to refactoring steps, including new binding steps.

Fields

hPrev, hNext :: ProcessStepID

descent :: StepInfo v x t -> StepInfo v x t

whatsHappen :: Ord a => a -> Process a i -> [Step a i]

extractInstructionAt :: (ProcessorUnit u v x t, Typeable u) => u -> t -> [Instruction u]

withShift :: (Eq a, Num a) => a -> a -> a

Shift nextTick value if it is not zero on a specific offset. Use case: The processor unit has buffered output, so we should provide oe signal for one tick before data actually send to the bus. That raises the following cases:

First usage. We can receive value immediately on nextTick

tick | Endpoint     | Instruction |
 0   | Target "c"   | WR          | <- nextTick

Not first usage. We need to wait for one tick from the last instruction due to the offset between instruction and data transfers.

tick | Endpoint     | Instruction |
  8  |              | OE          |
  9  | Source ["b"] |             | <- nextTick
 10  | Target "c"   | WR          |

isRefactorStep :: StepInfo v x t -> Bool

isAllocationStep :: StepInfo v x t -> Bool

Control

class Controllable pu where

Type class for controllable units. Defines two level of a unit behaviour representation (see ahead).

Associated Types

data Instruction pu :: Type

data Microcode pu :: Type

Microcode desctibe controll signals on each mUnit cycle (without exclusion).

Methods

zipSignalTagsAndValues :: Ports pu -> Microcode pu -> [(SignalTag, SignalValue)]

Zip port signal tags and value.

usedPortTags :: Ports pu -> [SignalTag]

Get list of used control signal tags.

takePortTags :: [SignalTag] -> pu -> Ports pu

Take signal tags from inifinite list of tags.

newtype SignalTag

Constructors

SignalTag

Fields

signalTag :: Text

class UnambiguouslyDecode pu where

Decoding microcode from a simple instruction (microcode don't change over time).

TODO: Generalize that class for all process units, including networks.

Methods

decodeInstruction :: Instruction pu -> Microcode pu

class Connected pu

Type class of processor units with control ports.

Associated Types

data Ports pu :: Type

A processor unit control ports (signals, flags).

class ByTime pu t | pu -> t where

Getting microcode value at a specific time.

Methods

microcodeAt :: pu -> t -> Microcode pu

data SignalValue

Control line value.

Constructors

Undef
undefined by design (x)

Bool Bool
boolean (`0` or `1`)

BrokenSignal
broken value (x) by data colision

(+++) :: SignalValue -> SignalValue -> SignalValue

IO

class IOConnected pu where

Type class of processor units with IO ports.

Minimal complete definition

Nothing

Associated Types

data IOPorts pu :: Type

Methods

inputPorts :: IOPorts pu -> Set InputPortTag

External input ports, which go outside of NITTA mUnit.

outputPorts :: IOPorts pu -> Set OutputPortTag

External output ports, which go outside of NITTA mUnit.

inoutPorts :: IOPorts pu -> Set InoutPortTag

External output ports, which go outside of NITTA mUnit.

newtype InputPortTag

Constructors

InputPortTag

Fields

inputPortTag :: Text

newtype OutputPortTag

Constructors

OutputPortTag

Fields

outputPortTag :: Text

newtype InoutPortTag

Constructors

InoutPortTag

Fields

inoutPortTag :: Text