NITTA.Model.Time

Description

Documentation

type VarValTime v x t = (Var v, Val x, Time t)

Shortcut for variable (v), value (x) and time (t) type constrains.

type VarValTimeJSON v x t = (VarValTime v x t, ToJSONKey v, ToJSON v, ToJSON x, ToJSON t)

type Time t = (Default t, Num t, Bounded t, Ord t, Show t, Typeable t, Enum t, Integral t)

Shortcut for time type constrain.

data TimeConstraint t

Time constrain for processor activity.

Constructors

TimeConstraint

Fields

tcAvailable :: Interval t

Inclusive interval, when value available to transfer.

tcDuration :: Interval t

Inclusive interval, possible for value transfers.

data TaggedTime tag t

Forgoten implementation of tagged time for speculative if statement. Current - dead code.

Constructors

TaggedTime

Fields

tTag :: Maybe tag

tClock :: t

Orphan instances

ToJSON t => ToJSON (Interval t)

Instance details

Methods

toJSON :: Interval t -> Value

toEncoding :: Interval t -> Encoding

toJSONList :: [Interval t] -> Value

toEncodingList :: [Interval t] -> Encoding