NITTA.Project.TestBench

Description

Documentation

class Testable m v x | m -> v x where

Type class for all testable parts of a target system.

Methods

testBenchImplementation :: Project m v x -> Implementation

class IOTestBench pu v x | pu -> v x where

Processor units with input/output ports should be tested by generation external input ports signals and checking output port signals.

Minimal complete definition

Nothing

Methods

testEnvironmentInitFlag :: Text -> pu -> Maybe Text

testEnvironment :: Text -> pu -> UnitEnv pu -> TestEnvironment v x -> Maybe Verilog

data TestEnvironment v x

Information required for testbench generation.

Constructors

TestEnvironment

Fields

teCntx :: Cntx v x

expected data

teComputationDuration :: Int

duration of computational process

data TestbenchReport v x

Constructors

TestbenchReport

Fields

tbStatus :: Bool

tbPath :: FilePath

tbFiles :: [FilePath]

tbFunctions :: [Text]

tbSynthesisSteps :: [Text]

tbCompilerDump :: Text

tbSimulationDump :: Text

tbFunctionalSimulationLog :: [HashMap v x]

tbLogicalSimulationLog :: [HashMap v x]

testBenchTopModuleName :: (TargetSystemComponent m, Testable m v x) => Project m v x -> FilePath

Get name of testbench top module.

verilogProjectFiles :: (TargetSystemComponent m, Testable m v x) => Project m v x -> [FilePath]

Generate list of project verilog files (including testbench).

data SnippetTestBenchConf m

Data Type for SnippetTestBench function

Constructors

SnippetTestBenchConf

Fields

tbcSignals :: [Text]

tbcPorts :: Ports m

tbcMC2verilogLiteral :: Microcode m -> Text

snippetTestBench :: forall m v x t. (WithFunctions m (F v x), ProcessorUnit m v x t, TargetSystemComponent m, UnambiguouslyDecode m, Typeable m, Show (Instruction m), Default (Microcode m)) => Project m v x -> SnippetTestBenchConf m -> Text

Function for testBench PU test