NITTA.Project.Types

Description

Documentation

data Project m v x

Target project for different purpose (testing, target system, etc). Should be writable to disk.

Constructors

Project

Fields

pName :: Text

target project name

pLibPath :: FilePath

IP-core library path

pTargetProjectPath :: FilePath

output path for target project

pAbsTargetProjectPath :: FilePath

absolute output path for target project

pInProjectNittaPath :: FilePath

relative to the project path output path for NITTA processor inside target project

pAbsNittaPath :: FilePath

absolute output path for NITTA processor inside target project

pUnit :: m

mUnit model (a mUnit unit for testbench or network for complete NITTA mUnit)

pUnitEnv :: UnitEnv m

pTestCntx :: Cntx v x

testbench context with input values

pTemplates :: [FilePath]

Target platform templates

defProjectTemplates :: [FilePath]

class TargetSystemComponent pu where

Type class for target components. Target -- a target system project or a testbench.

Methods

moduleName :: Text -> pu -> Text

Name of the structural hardware module or Verilog module name (network or process unit)

software :: Text -> pu -> Implementation

Software and other specification which depends on application algorithm

hardware :: Text -> pu -> Implementation

Hardware which depends on microarchitecture description and requires synthesis.

hardwareInstance :: Text -> pu -> UnitEnv pu -> Verilog

Generate code for making an instance of the hardware module

data Implementation

Element of target system implementation

Constructors

Immediate
Immediate implementation in the from of Ginger template (nitta.paths.nest + projectContext)

Fields

impFileName :: FilePath

impText :: Text

FromLibrary
Fetch implementation from library

Fields

impFileName :: FilePath

Aggregate
Aggregation of many implementation parts in separate paths

Fields

impPath :: Maybe FilePath

subComponents :: [Implementation]

Empty
Nothing

data UnitEnv m

Resolve uEnv element to verilog source code. E.g. dataIn into data_bus, dataOut into accum_data_out.

Constructors

UnitEnv

Fields

sigClk :: Text

clock signal

sigRst :: Text

reset signal

sigCycleBegin :: Text

posedge on computation cycle begin

sigInCycle :: Text

positive on computation cycle

sigCycleEnd :: Text

posedge on computation cycle end

ctrlPorts :: Maybe (Ports m)

ioPorts :: Maybe (IOPorts m)

valueIn, valueOut :: Maybe (Text, Text)

envInputPorts :: IOConnected pu => UnitEnv pu -> Set InputPortTag

envOutputPorts :: IOConnected pu => UnitEnv pu -> Set OutputPortTag

envInOutPorts :: IOConnected pu => UnitEnv pu -> Set InoutPortTag