NITTA.Synthesis

Description

TargetSynthesis is an entry point for synthesis process. TargetSynthesis flow shown on fig.

====================================================================================================================
                                                                                                             Prepare
NITTA.Synthesis:TargetSynthesis                                                                     NITTA.Project...
    # tName                                                                              NITTA.Frontends
    # tMicroArch -----------------------------
    # tSourceCode -+                          |     /--+-- mkModelWithOneNetwork
                   |                          |     |  |
                   *<-translate               |     |  |
                   |                          |     |  v      NITTA.Model:TargetSystem----------
    # tDFG +--------*-------- # mUnit         |             |    NITTA.Model...
        |                                        |                                |             |     /-----------
        |                                        v                                |             |     |  Target   |
        +----------------------------------------*-----------> # mDataFlowGraph   |             -----+  System   |
                                                                                  |                   | Imitation |
    # tReceivedValues                                                             |                   |   Model   |
    # tVerbose                                                                    |                   -----------/
    # tSynthesisMethod ----------------------------------                        |
                                                         |                        |
                                                         |                        |
===================================================================================================================
                                                         |                        |               Synthesis process
        NITTA.Synthesis.Types:Tree                       |                        |           NITTA.Synthesis.Types
            # sID                                        |                        |          NITTA.Synthesis.Method
            # sState <------------------------------------------------------------+
            # sDecision                                  |
                # option                                 |
                # decision                               *<-- search for best synthesis path
                # metrics                                |
                # score                                  v
            # sSubForestVar ------------------------->/ * / -----
                                                                   |
                                                                   v
                                                            NITTA.Synthesis.Types:Tree
                                                                # sState
                                                                    # sTarget
                                       /------------------------------- # mUnit
                                       |        /---------------------- # mDataFlowGraph
                                       |        |                   # ...
                                       |        |               # ...
                                       |        |
===================================================================================================================
                                       |        |                                         Target project generation
NITTA.Project.Types:Project            |        |                                                 NITTA.Project....
 |      # pName <--------- $tName      |        |
 |      # pLibPath                     |        +<----- $tReceivedValues
 |      # pTargetProjectPath           |        |
 |      # pModel<----------------------/        *<----- functional simulation (FIXME)
 |                                              |
 |      # pTestCntx <---------------------------/
 |
 |
 *<---------- $writeProject
 |                # TargetSystem
 |                    # hardware
 |                    # software
 |                # TestBench
 |                # Templates
 |
 ---> filesystem

Documentation

module NITTA.Synthesis.Explore

module NITTA.Synthesis.Method

module NITTA.Synthesis.Steps

module NITTA.Synthesis.Types

mkModelWithOneNetwork :: ProcessorUnit u v x t => u -> DataFlowGraph v x -> TargetSystem u tag v x t

Make a model of NITTA process with one network and a specific algorithm. All functions are already bound to the network.

data TargetSynthesis tag v x t

Description of synthesis task. Applicable for target system synthesis and testing purpose.

Constructors

TargetSynthesis

Fields

tName :: String

target name, used for top level module name and project path

tMicroArch :: BusNetwork tag v x t

composition of processor units, IO ports and its interconnect

tSourceCode :: Maybe Text

optional application source code (lua)

tDFG :: DataFlowGraph v x

algorithm in intermediate data flow graph representation (if tSourceCode present will be overwritten)

tReceivedValues :: [(v, [x])]

values from input interface for testing purpose

tSynthesisMethod :: SynthesisMethod tag v x t

synthesis method

tLibPath :: String

IP-core library directory

tPath :: String

output directory, where CAD create project directory with tName name

tTemplates :: [FilePath]

tSimulationCycleN :: Int

number of simulation and testbench cycles

tSourceCodeType :: FrontendType

source code format type

runTargetSynthesis :: (Bounded t, Suffix v1, Hashable v1, UnitTag tag, Val p, Default t, Show t, Integral t, Typeable t, Typeable v1, Ord v1, ToString v1, IsString v1) => TargetSynthesis tag v1 p t -> IO (Either String (TestbenchReport v1 p))

synthesizeTargetSystem :: (UnitTag tag, VarValTime v x t) => TargetSynthesis tag v x t -> IO (DefTree tag v x t, Either String (Project (BusNetwork tag v x t) v x))