NITTA.Synthesis.Explore

Description

Documentation

synthesisTreeRootIO :: (Val x, Hashable v, Suffix v, ToString v, ToString tag, IsString v, IsString tag, Bounded t, Typeable v, Typeable tag, Typeable t, Show t, Default t, Integral t, Ord v, Ord tag) => TargetSystem (BusNetwork tag v x t) tag v x t -> IO (Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t)

Make synthesis tree

getTreeIO :: (IsString tag, IsString v, ToString tag, ToString v, Suffix v, Hashable v, Val x, Bounded t, Typeable tag, Typeable v, Typeable t, Show t, Default t, Integral t, Ord tag, Ord v) => Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> SID -> IO (Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t)

Get specific by nId node from a synthesis tree.

getTreePathIO :: (IsString tag, IsString v, ToString tag, ToString v, Suffix v, Hashable v, Val x, Bounded t, Typeable tag, Typeable v, Typeable t, Show t, Default t, Integral t, Ord tag, Ord v) => Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> SID -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

Get list of all nodes from root to selected.

subForestIO :: (IsString tag, IsString v, ToString tag, ToString v, Suffix v, Hashable v, Val x, Bounded t, Typeable tag, Typeable v, Typeable t, Show t, Default t, Integral t, Ord tag, Ord v) => Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

Get all available edges for the node. Edges calculated only for the first call.

positiveSubForestIO :: (IsString tag, IsString v, ToString tag, ToString v, Suffix v, Hashable v, Val x, Bounded t, Typeable tag, Typeable v, Typeable t, Show t, Default t, Integral t, Ord tag, Ord v) => Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

For synthesis method is more usefull, because throw away all useless trees in subForest (objective function value less than zero).

isComplete :: ProcessorUnit u v1 x1 t1 => Tree (TargetSystem u tag1 v1 x1 t1) tag2 v2 x2 t2 -> Bool

isLeaf :: Tree m tag v x t -> Bool