NITTA.Synthesis.Method

Description

Documentation

type DefTree tag v x t = Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t

Default synthesis tree type.

type SynthesisMethod tag v x t = DefTree tag v x t -> IO (DefTree tag v x t)

The synthesis method is a function, which manipulates a synthesis tree. It receives a node and explores it deeply by IO.

simpleSynthesisIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t

Schedule process by simple synthesis.

smartBindSynthesisIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t

obviousBindThreadIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t

allBestThreadIO :: (VarValTime v x t, UnitTag tag) => Int -> SynthesisMethod tag v x t

stateOfTheArtSynthesisIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t

The most complex synthesis method, which embedded all another. That all.

allBindsAndRefsIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t

bestStepIO :: (VarValTime v x t, UnitTag tag) => SynthesisMethod tag v x t