NITTA.Synthesis.MlBackend.Client

Description

Documentation

data ScoringInput tag v x t

Constructors

ScoringInput

Fields

scoringTarget :: ScoringTarget

nodes :: [NodeView tag v x t]

data ScoringTarget

Constructors

ScoringTargetSid Sid

ScoringTargetAll

predictScoresIO :: forall {m} {v} {x} {t} {a} {tag}. (IsString v, Suffix v, Hashable v, Val x, Bounded t, ToJSONKey v, MonadThrow m, ToString v, ToString a, MonadIO m, Typeable v, Typeable t, Show t, Default t, Integral t, Ord v, ToJSON v, ToJSON t, ToJSON tag) => Text -> a -> [ScoringInput tag v x t] -> m [[Float]]