NITTA.Synthesis.MlBackend.FixedCache

Description

Taken as is from Control.Concurrent.Cache of io-memoize package, but with expoted Cache constructor, so we can use it in pattern matching.

This was necessary not to rewrite similar functionality. Linter errors were fixed as well.

Documentation

newtype Cache a

A thread-safe write-once cache. If you need more functionality, (e.g. multiple write, cache clearing) use an MVar instead.

Constructors

Cache (MVar (Maybe a))

newCache :: IO (Cache a)

Create an empty cache.

fetch :: Cache a -> IO a -> IO a

Fetch the value stored in the cache, or call the supplied fallback and store the result, if the cache is empty.