NITTA.Synthesis.Steps.Bind

Description

Documentation

data BindMetrics

Constructors

SingleBindMetrics

Fields

pCritical :: Bool

Can this binding block another one (for example, one Loop can take the last free buffer)?

pAlternative :: Float

How many alternative binding we have?

pRestless :: Float

How many ticks requires for executing the function?

pOutputNumber :: Float

pAllowDataFlow :: Float

How many transactions can be executed with this function?

pPossibleDeadlock :: Bool

May this binding cause deadlock?

pNumberOfBoundFunctions :: Float

pPercentOfBoundInputs :: Float

number of bound input variables / number of all input variables

pWave :: Maybe Float

GroupBindMetrics

Fields

pOnlyObviousBinds :: Bool

We don't have alternatives for binding

pFunctionPercentInBinds :: Float

number of bound functions / number of all functions in DFG

pAvgBinds :: Float

average number of binds per unit

pVarianceBinds :: Float

variance of binds per unit

pAvgUnitWorkload :: Float

average number of variables after bind per unit

pVarianceUnitWorkload :: Float

variance of variables after bind per unit

isSingleBind :: SynthesisDecision ctx m -> Bool

isMultiBind :: SynthesisDecision ctx m -> Bool

isObviousMultiBind :: SynthesisDecision ctx m -> Bool

Orphan instances

(UnitTag tag, VarValTime v x t) => SynthesisDecisionCls (SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t) (TargetSystem (BusNetwork tag v x t) tag v x t) (Bind tag v x) (Bind tag v x) BindMetrics

Instance details

Methods

decisions :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> Bind tag v x -> [(Bind tag v x, TargetSystem (BusNetwork tag v x t) tag v x t)]

parameters :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> Bind tag v x -> Bind tag v x -> BindMetrics

estimate :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> Bind tag v x -> Bind tag v x -> BindMetrics -> Float