NITTA.Synthesis.Steps.Dataflow

Description

Documentation

data DataflowMetrics

Constructors

DataflowMetrics

Fields

pWaitTime :: Float

pRestrictedTime :: Bool

pNotTransferableInputs :: [Float]

number of variables, which is not transferable for affected functions.

pFirstWaveOfTargetUse :: Float

number of the first wave in which one of the target variables is used

Orphan instances

(UnitTag tag, VarValTime v x t) => SynthesisDecisionCls (SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t) (TargetSystem (BusNetwork tag v x t) tag v x t) (DataflowSt tag v (TimeConstraint t)) (DataflowSt tag v (Interval t)) DataflowMetrics

Instance details

Methods

decisions :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> DataflowSt tag v (TimeConstraint t) -> [(DataflowSt tag v (Interval t), TargetSystem (BusNetwork tag v x t) tag v x t)]

parameters :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> DataflowSt tag v (TimeConstraint t) -> DataflowSt tag v (Interval t) -> DataflowMetrics

estimate :: SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> DataflowSt tag v (TimeConstraint t) -> DataflowSt tag v (Interval t) -> DataflowMetrics -> Float