NITTA.Synthesis.Types

Description

Synthesis can be represented as a graph (tree), where each Node describes the target system ModelState and each Edge synthesis decision.

A synthesis graph is very large and calculating and storing it in memory is very bad idea. Also, working with synthesis graph usually making from the specific node, not from the root. As a result, synthesis graph design as a explicit lazy mutable structure implemented by TVar.

From this point of view, the synthesis process is a finding of the best tree leaf (lowest process duration for finished synthesis), and the best synthesis method - a method which directly walks over the tree to the best leaf without wrong steps.

Documentation

class SynthesisDecisionCls ctx m o d p | ctx o -> m d p where

Methods

decisions :: ctx -> o -> [(d, m)]

parameters :: ctx -> o -> d -> p

estimate :: ctx -> o -> d -> p -> Float

data Tree m tag v x t

Synthesis tree

Constructors

Tree

Fields

sID :: Sid

sState :: SynthesisState m tag v x t

sDecision :: SynthesisDecision (SynthesisState m tag v x t) m

sSubForestVar :: TMVar [Tree m tag v x t]

lazy mutable field with different synthesis options and sub nodes

isLeaf :: Bool

isComplete :: Bool

data SynthesisDecision ctx m where

Constructors

Root :: SynthesisDecision ctx m

SynthesisDecision

Fields

:: (Typeable p, SynthesisDecisionCls ctx m o d p, Show d, ToJSON p, Viewable d DecisionView)

=> { option :: o

, decision :: d

, metrics :: p

, scores :: Map Text Float

} -> SynthesisDecision ctx m

data SynthesisState m tag v x t

Constructors

SynthesisState

Fields

sParent :: Maybe (Tree m tag v x t)

sTarget :: m

sAllocationOptions :: [Allocation tag]

PU allocation options cache

sBindOptions :: [Bind tag v x]

bind options cache

sResolveDeadlockOptions :: [ResolveDeadlock v x]

sOptimizeAccumOptions :: [OptimizeAccum v x]

sOptimizeLogicalUnitOptions :: [OptimizeLogicalUnit v x]

sConstantFoldingOptions :: [ConstantFolding v x]

sBreakLoopOptions :: [BreakLoop v x]

sDataflowOptions :: [DataflowSt tag v (TimeConstraint t)]

dataflow options cache

bindingAlternative :: Map (F v x) [tag]

a map from functions to possible processor unit tags

possibleDeadlockBinds :: Set (F v x)

a function set, which binding may cause dead lock

bindWaves :: Map v Int

if algorithm will be represented as a graph, where nodes - variables of not bound functions, edges - casuality, wave is a minimal number of a step from an initial node to selected

processWaves :: [ProcessWave v x]

Execution waves of the algorithm. See detailed description in NITTA.Intermediate.Analysis module.

numberOfProcessWaves :: Int

Number of execution waves of the algorithm.

numberOfDataflowOptions :: Int

number of dataflow options

transferableVars :: Set v

a variable set, which can be transferred on the current synthesis step

unitWorkloadInFunction :: Map tag Int

dictionary with number of bound functions for each unit

type SynthesisMethodConstraints tag v x t = (VarValTimeJSON v x t, ToJSON tag, UnitTag tag)

Shortcut for constraints in signatures of synthesis method functions. This used to be (VarValTime v x t, UnitTag tag). See below for more info.

newtype Sid

Synthesis node ID. ID is a relative path, encoded as a sequence of an option index.

Constructors

Sid [Int]

type DefTree tag v x t = Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t

Default synthesis tree type.

type SynthesisMethod tag v x t = DefTree tag v x t -> IO (DefTree tag v x t)

The synthesis method is a function, which manipulates a synthesis tree. It receives a node and explores it deeply by IO.

(<?>) :: Num p => Bool -> p -> p

targetUnit :: Tree (TargetSystem c tag v x t) tag v x t -> c

targetDFG :: Tree (TargetSystem u tag v x t) tag v x t -> DataFlowGraph v x

defScore :: SynthesisDecision ctx m -> Float

mlScoreKeyPrefix :: IsString a => a