NITTA.UIBackend

Description

Documentation

backendServer :: (Suffix w, Hashable w, Val w, Bounded w, ToJSONKey w, FromHttpApiData w, UnitTag w, Label w, ToString w, IsString w, Typeable w, Typeable w, Show w, Default w, Integral w, Ord w, ToJSON w, ToJSON w, ToJSON w) => Int -> BackendCtx w w w w -> IO ()

Run backend server.

prepareJSAPI :: Interpolatable 'True p Text => p -> FilePath -> IO ()

restDocs :: Show a => a -> String

apiPath :: FilePath