NITTA.UIBackend.REST

Description

Documentation

type SynthesisAPI tag v x t = (Description "Get whole synthesis tree" :> ("synthesisTree" :> Get '[JSON] (TreeView ShortNodeView))) :<|> ((Description "Get synthesis tree info" :> ("treeInfo" :> Get '[JSON] TreeInfo)) :<|> ("node" :> (Capture "sid" SID :> (SynthesisTreeNavigationAPI tag v x t :<|> (NodeInspectionAPI tag v x t :<|> (TestBenchAPI v x :<|> (SynthesisMethodsAPI tag v x t :<|> SynthesisPracticesAPI tag v x t)))))))

synthesisServer :: (IsString tag, IsString v, ToString tag, ToString v, Suffix v, Hashable v, Val x, Bounded a, ToJSONKey v, ToJSON v, ToJSON x, ToJSON a, Label v, MonadIO m1, MonadIO m2, MonadIO m3, MonadIO m4, MonadIO m5, MonadIO m6, MonadIO m7, MonadIO m8, MonadIO m9, MonadIO m10, MonadIO m11, MonadIO m12, MonadIO m13, MonadIO m14, MonadIO m15, MonadIO m16, MonadIO m17, MonadIO m18, MonadIO m19, MonadIO m20, MonadIO m21, Typeable tag, Typeable v, Typeable a, Show a, Default a, Integral a, Ord tag, Ord v) => BackendCtx tag v x a -> m2 (TreeView ShortNodeView) :<|> (m1 TreeInfo :<|> (SID -> (m4 [NodeView tag v x a] :<|> (m5 (Maybe (NodeView tag v x a)) :<|> m3 [NodeView tag v x a])) :<|> ((m8 (NodeView tag v x a) :<|> (m7 (GraphStructure GraphEdge) :<|> (m6 (ProcessTimelines a) :<|> (m9 (Process a StepInfoView) :<|> ((tag -> m10 (Process a StepInfoView)) :<|> (m11 [UnitEndpoints tag v a] :<|> (m12 (MicroarchitectureDesc tag) :<|> m13 (Debug tag v a)))))))) :<|> ((String -> Int -> m14 (TestbenchReport v x)) :<|> ((m15 SID :<|> (m16 SID :<|> m17 SID)) :<|> (m18 SID :<|> (m19 SID :<|> (m20 SID :<|> (Int -> m21 SID)))))))))

data BackendCtx tag v x t

Constructors

BackendCtx

Fields

root :: DefTree tag v x t

root synthesis node

receivedValues :: [(v, [x])]

lists of received by IO values

outputPath :: String

data UnitEndpoints tag v t

Constructors

UnitEndpoints

Fields

unitTag :: tag

unitEndpoints :: [EndpointSt v (TimeConstraint t)]

Orphan instances

ToSample Char

Instance details

Methods

toSamples :: Proxy Char -> [(Text, Char)]

ToSample SID

Instance details

Methods

toSamples :: Proxy SID -> [(Text, SID)]

UnitTag tag => ToSample [(Text, [Lock tag])]

Instance details

Methods

toSamples :: Proxy [(Text, [Lock tag])] -> [(Text, [(Text, [Lock tag])])]

UnitTag tag => ToSample (Lock tag)

Instance details

Methods

toSamples :: Proxy (Lock tag) -> [(Text, Lock tag)]

UnitTag tag => ToSample (MicroarchitectureDesc tag)

Instance details

Methods

toSamples :: Proxy (MicroarchitectureDesc tag) -> [(Text, MicroarchitectureDesc tag)]

ToCapture (Capture "sid" SID)

Instance details

Methods

toCapture :: Proxy (Capture "sid" SID) -> DocCapture

ToCapture (Capture "tag" tag :: Type)

Instance details

Methods

toCapture :: Proxy (Capture "tag" tag) -> DocCapture

ToParam (QueryParam' mods "loopsNumber" Int :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "loopsNumber" Int) -> DocQueryParam

ToParam (QueryParam' mods "pName" String :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "pName" String) -> DocQueryParam

ToParam (QueryParam' mods "deep" Int :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "deep" Int) -> DocQueryParam

ToSample (EndpointSt String (TimeConstraint Int))

Instance details

Methods

toSamples :: Proxy (EndpointSt String (TimeConstraint Int)) -> [(Text, EndpointSt String (TimeConstraint Int))]

Time t => ToSample (Process t StepInfoView)

Instance details

Methods

toSamples :: Proxy (Process t StepInfoView) -> [(Text, Process t StepInfoView)]