NITTA.UIBackend.ViewHelper

Description

We can not autogenerate ToJSON implementation for some types, so we add helper types for doing that automatically. Why do we need to generate ToJSON automatically? We don't want to achieve consistency between client and server manually.

Documentation

module NITTA.UIBackend.ViewHelperCls

module NITTA.Model.Problems.ViewHelper

data FView

Helper for JSON serialization

Constructors

FView

Fields

fvFun :: Text

fvHistory :: [Text]

class Viewable t v | t -> v where

Type class of helpers required for autogeneration of ToJSON and typescript types.

Methods

view :: t -> v

viewNodeTree :: (Enum b, Num b, NextTick u b) => Tree (TargetSystem u tag v x t) tag v x t -> IO (TreeView ShortNodeView)

data TreeView a

data ShortNodeView

data NodeView tag v x t

newtype StepInfoView

Constructors

StepInfoView Text

type VarValTimeJSON v x t = (VarValTime v x t, ToJSONKey v, ToJSON v, ToJSON x, ToJSON t)

Orphan instances

ToJSON TreeInfo

Instance details

Methods

toJSON :: TreeInfo -> Value

toEncoding :: TreeInfo -> Encoding

toJSONList :: [TreeInfo] -> Value

toEncodingList :: [TreeInfo] -> Encoding

ToSample TreeInfo

Instance details

Methods

toSamples :: Proxy TreeInfo -> [(Text, TreeInfo)]

ToSample Integer

Instance details

Methods

toSamples :: Proxy Integer -> [(Text, Integer)]

(ToJSONKey v, ToJSON v, ToJSON x) => ToJSON (TestbenchReport v x)

Instance details

Methods

toJSON :: TestbenchReport v x -> Value

toEncoding :: TestbenchReport v x -> Encoding

toJSONList :: [TestbenchReport v x] -> Value

toEncodingList :: [TestbenchReport v x] -> Encoding

ToSample (TestbenchReport String Int)

Instance details

Methods

toSamples :: Proxy (TestbenchReport String Int) -> [(Text, TestbenchReport String Int)]

(Var v, Time t) => Viewable (Process t (StepInfo v x t)) (Process t StepInfoView)

Instance details

Methods

view :: Process t (StepInfo v x t) -> Process t StepInfoView

(Var v, Time t) => Viewable (StepInfo v x t) StepInfoView

Instance details

Methods

view :: StepInfo v x t -> StepInfoView

(UnitTag tag, VarValTimeJSON v x t) => Viewable (DefTree tag v x t) (NodeView tag v x t)

Instance details

Methods

view :: DefTree tag v x t -> NodeView tag v x t