NITTA.Utils.ProcessDescription

Description

A multilevel process is an object with a complex internal structure. Process description should contain every step (including start and finish time), and relations between them (Vertical or sequence). It is possible to define process manually, but, in practice, is preferred to use State based builder from that module.

It also agreed to the process inspection.

Documentation

runSchedule :: ProcessorUnit u v x t => u -> State (Schedule u v x t) a -> (a, Process t (StepInfo v x t))

Execute process builder and return list of new step UID and new process description. The initial process state is getting from the PU by the process function.

execSchedule :: ProcessorUnit u v x t => u -> State (Schedule u v x t) a -> Process t (StepInfo v x t)

Execute process builder and return new process description. The initial process state is getting from the PU by the process function.

execScheduleWithProcess :: pu -> Process t (StepInfo v x t) -> State (Schedule pu v x t) a -> Process t (StepInfo v x t)

Execute process builder and return new process description. The initial process state is passed explicetly.

Why can not we get a process here? In the case of Bus Network, it also fetches processes from underlying units.

scheduleStep :: MonadState (Schedule pu v x t) m => Interval t -> StepInfo v x t -> m [ProcessStepID]

Add process step with passed the time and info.

scheduleEndpoint :: forall {m} {pu} {v} {x} {t}. MonadState (Schedule pu v x t) m => EndpointSt v (Interval t) -> m [ProcessStepID] -> m [ProcessStepID]

Add to the process description information about endpoint behaviour, and it's low-level implementation (on instruction level). Vertical relations connect endpoint level and instruction level steps.

scheduleEndpoint_ :: forall {f} {pu} {v} {x} {t}. MonadState (Schedule pu v x t) f => EndpointSt v (Interval t) -> f [ProcessStepID] -> f ()

scheduleFunctionBind :: (MonadState (Schedule pu v x t) m, Show a) => a -> m [ProcessStepID]

scheduleFunctionRevoke :: (MonadState (Schedule pu v x t) m, Show a) => a -> m [ProcessStepID]

scheduleFunction :: MonadState (Schedule pu v x t) m => Interval t -> F v x -> m [ProcessStepID]

Add to the process description information about function evaluation.

scheduleFunctionFinish :: (Ord v, Typeable v, IsString v, ToString v, Suffix v, Hashable v) => [ProcessStepID] -> F v x -> Interval t -> StateT (Schedule pu v x t) Identity [ProcessStepID]

Schedule function and establish vertical relations between bind step, function step, and all related endpoints.

scheduleFunctionFinish_ :: (Ord v, Typeable v, IsString v, ToString v, Suffix v, Hashable v) => [ProcessStepID] -> F v x -> Interval t -> StateT (Schedule pu v x t) Identity ()

scheduleRefactoring :: (MonadState (Schedule pu v x t) m, Typeable ref, Show ref, Eq ref) => Interval t -> ref -> m [ProcessStepID]

scheduleInstructionUnsafe :: (MonadState (Schedule pu v x t) m, Show (Instruction pu), Typeable pu, Num t) => Interval t -> Instruction pu -> m [ProcessStepID]

Add to the process description information about instruction evaluation. Unsafe means: without instruction collision check and nextTick consistency.

scheduleInstructionUnsafe_ :: forall {f} {pu} {v} {x} {t}. (MonadState (Schedule pu v x t) f, Show (Instruction pu), Typeable pu, Num t) => Interval t -> Instruction pu -> f ()

scheduleNestedStep :: forall {m} {pu} {v} {x} {t} {tag}. (MonadState (Schedule pu v x t) m, UnitTag tag) => tag -> Step t (StepInfo v x t) -> m ProcessStepID

Add to the process description information about nested step.

establishVerticalRelations :: forall {m} {pu} {v} {x} {t}. MonadState (Schedule pu v x t) m => [ProcessStepID] -> [ProcessStepID] -> m ()

Add to the process description information about vertical relations, which are defined by the Cartesian product of high and low lists.

establishHorizontalRelations :: forall {m} {pu} {v} {x} {t}. MonadState (Schedule pu v x t) m => [ProcessStepID] -> [ProcessStepID] -> m ()

Add to the process description information about horizontal relations (inside level), which are defined by the Cartesian product of high and low lists.

getProcessSlice :: State (Schedule pu v x t) (Process t (StepInfo v x t))

Get a current slice of the computational process.

relatedEndpoints :: Ord a => Process t (StepInfo a x t) -> Set a -> [Step t (StepInfo a x t)]

castInstruction :: (Typeable a, Typeable pu) => pu -> a -> Maybe (Instruction pu)

Helper for instruction extraction from a rigid type variable.

scheduleAllocation :: (MonadState (Schedule pu v x t) m, Typeable a, Show a, Eq a) => a -> m [ProcessStepID]