NITTA.Synthesis.Explore

Description

Documentation

synthesisTreeRootIO :: (UnitTag tag, Val x, Bounded t, Hashable v, Suffix v, Show t, Default t, Integral t, Typeable t, Typeable v, Ord v, ToString v, IsString v) => TargetSystem (BusNetwork tag v x t) tag v x t -> IO (Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t)

Make synthesis tree

getTreeIO :: (UnitTag tag, Hashable v, Suffix v, Val x, Bounded t, ToJSONKey v, Show t, Default t, Integral t, ToJSON v, ToJSON t, ToJSON tag, Typeable v, Typeable t, Ord v, ToString v, IsString v) => BackendCtx tag v x t -> Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> Sid -> IO (Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t)

Get specific by nId node from a synthesis tree.

getTreePathIO :: (UnitTag tag, Hashable v, Suffix v, Val x, Bounded t, ToJSONKey v, Show t, Default t, Integral t, ToJSON v, ToJSON t, ToJSON tag, Typeable v, Typeable t, Ord v, ToString v, IsString v) => BackendCtx tag v x t -> Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> Sid -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

Get list of all nodes from root to selected.

subForestIO :: (UnitTag tag, Hashable v, Suffix v, Val x, Bounded t, ToJSONKey v, ToJSON v, ToJSON t, ToJSON tag, Show t, Default t, Integral t, Typeable v, Typeable t, Ord v, ToString v, IsString v) => BackendCtx tag v x t -> Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

Get all available edges for the node. Edges calculated only for the first call.

positiveSubForestIO :: (UnitTag tag, Hashable v, Suffix v, Val x, Bounded t, ToJSONKey v, Show t, Default t, Integral t, ToJSON v, ToJSON t, ToJSON tag, Typeable v, Typeable t, Ord v, ToString v, IsString v) => BackendCtx tag v x t -> Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t -> IO [Tree (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t]

For synthesis method is more usefull, because throw away all useless trees in subForest (objective function value less than zero).