NITTA.UIBackend.REST

Description

Documentation

type SynthesisAPI tag v x t = (Description "Get whole synthesis tree" :> ("synthesisTree" :> Get '[JSON] (TreeView ShortNodeView))) :<|> ((Description "Get synthesis tree info" :> ("treeInfo" :> Get '[JSON] TreeInfo)) :<|> ("node" :> (Capture "sid" Sid :> (SynthesisTreeNavigationAPI tag v x t :<|> (NodeInspectionAPI tag v x t :<|> (TestBenchAPI v x :<|> (SynthesisMethodsAPI tag v x t :<|> SynthesisPracticesAPI tag v x t)))))))

synthesisServer :: (UnitTag tag2, Bounded b, Hashable v, Suffix v, Val x, ToJSONKey v, Label v, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, MonadIO m, Show b, Default b, Integral b, ToJSON v, ToJSON b, ToJSON tag2, Typeable b, Typeable v, Ord v, ToString v, IsString v) => BackendCtx tag2 v x b -> m (TreeView ShortNodeView) :<|> (m TreeInfo :<|> (Sid -> (m [NodeView tag2 v x b] :<|> (m (Maybe (NodeView tag2 v x b)) :<|> m [NodeView tag2 v x b])) :<|> ((m (NodeView tag2 v x b) :<|> (m (GraphStructure GraphEdge) :<|> (m (ProcessTimelines b) :<|> (m (Process b StepInfoView) :<|> ((tag2 -> m (Process b StepInfoView)) :<|> (m [UnitEndpoints tag2 v b] :<|> (m (MicroarchitectureDesc tag2) :<|> m (Debug tag2 v b)))))))) :<|> ((String -> Int -> m (TestbenchReport v x)) :<|> ((m Sid :<|> (m Sid :<|> m Sid)) :<|> (m Sid :<|> (m Sid :<|> (m Sid :<|> (Int -> m Sid)))))))))

data BackendCtx tag v x t

Constructors

BackendCtx

Fields

root :: DefTree tag v x t

root synthesis node

receivedValues :: [(v, [x])]

lists of received by IO values

outputPath :: String

mlBackendGetter :: IO MlBackendServer

nodeScores :: [Text]

list of node scores name. Used for ML backend mostly, but can be used for any "hardcoded" scoring function.

data UnitEndpoints tag v t

Constructors

UnitEndpoints

Fields

unitTag :: tag

unitEndpoints :: [EndpointSt v (TimeConstraint t)]

Orphan instances

ToSample Sid

Instance details

Methods

toSamples :: Proxy Sid -> [(Text, Sid)]

ToSample Char

Instance details

Methods

toSamples :: Proxy Char -> [(Text, Char)]

ToCapture (Capture "sid" Sid)

Instance details

Methods

toCapture :: Proxy (Capture "sid" Sid) -> DocCapture

ToCapture (Capture "tag" tag :: Type)

Instance details

Methods

toCapture :: Proxy (Capture "tag" tag) -> DocCapture

ToParam (QueryParam' mods "deep" Int :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "deep" Int) -> DocQueryParam

ToParam (QueryParam' mods "loopsNumber" Int :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "loopsNumber" Int) -> DocQueryParam

ToParam (QueryParam' mods "pName" String :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "pName" String) -> DocQueryParam

UnitTag tag => ToSample (Lock tag)

Instance details

Methods

toSamples :: Proxy (Lock tag) -> [(Text, Lock tag)]

UnitTag tag => ToSample (MicroarchitectureDesc tag)

Instance details

Methods

toSamples :: Proxy (MicroarchitectureDesc tag) -> [(Text, MicroarchitectureDesc tag)]

UnitTag tag => ToSample [(Text, [Lock tag])]

Instance details

Methods

toSamples :: Proxy [(Text, [Lock tag])] -> [(Text, [(Text, [Lock tag])])]

ToSample (EndpointSt String (TimeConstraint Int))

Instance details

Methods

toSamples :: Proxy (EndpointSt String (TimeConstraint Int)) -> [(Text, EndpointSt String (TimeConstraint Int))]

Time t => ToSample (Process t StepInfoView)

Instance details

Methods

toSamples :: Proxy (Process t StepInfoView) -> [(Text, Process t StepInfoView)]